Update and rename csharp/NUnit/W3CChromeTest.cs to csharp/NUnitExamples.cs
Merging the W3CFirefoxTest class into this class and renaming for all NUnit Examples. This class utilizes the [OneTimeSetUp] attribute to obtain the username and accessKey from the environment variables. It uses [SetUp] attribute for setting up the SauceOptions. This could be moved into each test method, but it will reduce code duplication by being in the SetUp method.